thread: use unsigned (long) instead of size_t in tls_align test

+++ src/thread/tls_align.c
@@ -1,11 +1,10 @@
-#include <stddef.h>
 #include "test.h"
 
 extern struct {
 	char *name;
-	size_t size;
-	size_t align;
-	size_t addr;
+	unsigned size;
+	unsigned align;
+	unsigned long addr;
 } t[4];
 
 int main()
@@ -17,7 +16,7 @@ int main()
 			error("name is not set for t[%d]\n", i);
 		if (t[i].addr & (t[i].align-1))
 			error("bad alignment: %s, size: %u, align: %u, addr: 0x%lx\n",
-				t[i].name, (unsigned)t[i].size, (unsigned)t[i].align, (unsigned long)t[i].addr);
+				t[i].name, t[i].size, t[i].align, t[i].addr);
 	}
 	return test_status;
 }
